expose default typegen settings
